requirements_base.txt
otree-core>=1.4.29

# Heroku requires Django to be explicitly in requirements file
# in order to run collectstatic
Django==1.8.8
